feat: :sparkles: move AppleSiliconChip to a powermetrics file
so it can handle both CPU and GPU
